Refactor to controller, remove
